+++ ios/Classes/MusicFeatureAnalyzerPlugin.swift
@@ -32,7 +32,7 @@ public class MusicFeatureAnalyzerPlugin: NSObject, FlutterPlugin {
                 "loadingMode": "IOS_PLATFORM",
                 "handlerSet": channel != nil,
                 "platform": "iOS",
-                "pluginVersion": "1.0.2"
+                "pluginVersion": "1.0.3"
             ]
             result(response)
             return